checks fix

+++ Sources/Map.cpp
@@ -1,4 +1,4 @@
-
+// Copyright 2019 CERZAR
 #include "../Headers/Map.h"
 
 void Map::init_map_vector() {}
@@ -88,8 +88,8 @@ void Map::createMap() {
   startX += tileSize;
   startY += tileSize;
   endX =
-      vector[(int)vector.size() - 1]->rect.getPosition().x -
-      tileSize;
+      static_cast<int>(vector[(int)vector.size() - 1]->rect.getPosition().x -
+      tileSize);
   endY = static_cast<int>(vector[vector.size() - 1]->rect.getPosition().y) -
          tileSize;
 }
@@ -142,7 +142,7 @@ void Map::checkPlayerCollision(Object *target) {
               vector[counter]->rect.getGlobalBounds().left +
                   vector[counter]->rect.getGlobalBounds().width,
               target->rect.getGlobalBounds().top);
-        else if (target->direction == -3 || target->direction == -4) 
+        else if (target->direction == -3 || target->direction == -4)
         {
           target->rect.setPosition(target->rect.getGlobalBounds().left,
                                    vector[counter]->rect.getGlobalBounds().top -
@@ -179,7 +179,7 @@ void Map::checkPlayerCollision(Object *target) {
               vector[counter]->rect.getGlobalBounds().left +
                   vector[counter]->rect.getGlobalBounds().width,
               target->rect.getGlobalBounds().top);
-        else if (target->direction == -1 || target->direction == -2) 
+        else if (target->direction == -1 || target->direction == -2)
         {
           target->rect.setPosition(
               target->rect.getGlobalBounds().left,
@@ -262,14 +262,16 @@ void Map::checkBulletCollision(Object *target) {
   for (iter = vector.begin(); iter != vector.end(); iter++) {
     if (target->rect.getGlobalBounds().intersects(
             vector[counter]->rect.getGlobalBounds())) {
-      if (vector[counter]->type == BOX) {
+      if (vector[counter]->type == BOX)
+      {
         vector[counter]->hp--;
         target->isExist = false;
-        if (vector[counter]->hp <= 0) 
+        if (vector[counter]->hp <= 0)
         {
           vector[counter]->isExist = false;
         }
-      } else if (vector[counter]->type == WALL)
+      }
+      else if (vector[counter]->type == WALL)
         target->isExist = false;
     }
     counter++;
